Tweaks

+++ src/model/runningtime.cpp
@@ -3,13 +3,13 @@ using namespace std::chrono;
 
 static auto startupTime = system_clock::now();
 
-unsigned runningTimeMs()
+int64_t runningTimeMs()
 {
     const auto dT = duration_cast<milliseconds>(system_clock::now() - startupTime);
     return dT.count();
 }
 
-unsigned runningTimeS()
+int64_t runningTimeS()
 {
     const auto dT = duration_cast<seconds>(system_clock::now() - startupTime);
     return dT.count();
