+++ src/model/runningtime.h
@@ -1,8 +1,9 @@
 #ifndef RUNNINGTIME_H
 #define RUNNINGTIME_H
 
+#include <stdint.h>
 
-unsigned runningTimeMs();
-unsigned runningTimeS();
+int64_t runningTimeMs();
+int64_t runningTimeS();
 
 #endif // RUNNINGTIME_H
